Modified the header.ejs files so the menu will adapt to the HOST url address by using process.env.HOST in the href instead of hard coding the url path. This means changes can be made and pushed up to github and the live server or test server without having to change the href links to suit the HOST url.
Needed to install npm mime as I was getting an error after re-installing all of the node dependencies. Not sure why this happened but installing npm mime fixed it.

changed the git ignore to include the changed header file and _config file as they should now be agnostic to the the HOST url because the HOST is now set in the .env file, which still needs to customized to each HOST, whether local host live server or test server.